+++ src/components/molcule/carousel/index.tsx
@@ -1,5 +1,4 @@
 import * as S from './style';
-
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
