+++ src/components/molcule/list-tag/index.tsx
@@ -1,8 +1,8 @@
 import { categories } from 'data/shared';
-import { useEffect, useState } from 'react';
 import { BoxTag, Tag } from 'components/index';
 import { useFormDataActions } from 'store/formData';
 import { useActiveCategory, useProductListActions } from 'store/productListData';
+import { useMouseEventHandler } from 'hooks/useMouseEventHandler';
 
 interface ListTagProps {
   isform?: boolean;
@@ -10,34 +10,11 @@ interface ListTagProps {
 
 const ListTag = ({ isform }: ListTagProps) => {
   const { setFormData } = useFormDataActions();
-
   const activeCategory = useActiveCategory();
   const { setActiveCategory } = useProductListActions();
-
-  const [isDragging, setIsDragging] = useState(false);
-  const [startX, setStartX] = useState(0);
-  const [scrollLeft, setScrollLeft] = useState(0);
+  const { startDragging, stopDragging, whileDragging } = useMouseEventHandler();
 
   const list = isform ? categories.slice(1) : categories;
-
-  // 마우스 이벤트 핸들러
-  const startDragging = (e: React.PointerEvent<HTMLDivElement>) => {
-    setIsDragging(true);
-    setStartX(e.pageX - e.currentTarget.offsetLeft);
-    setScrollLeft(e.currentTarget.scrollLeft);
-  };
-
-  const stopDragging = () => {
-    setIsDragging(false);
-  };
-
-  const whileDragging = (e: React.PointerEvent<HTMLDivElement>) => {
-    if (!isDragging) return;
-    e.preventDefault();
-    const x = e.pageX - e.currentTarget.offsetLeft;
-    const walk = (x - startX) * 1;
-    e.currentTarget.scrollLeft = scrollLeft - walk;
-  };
 
   const handleClick = (category: string) => {
     if (isform) {
