+++ src/hooks/useMouseEventHandler.ts
@@ -1,0 +1,27 @@
+import { useState } from 'react';
+
+export const useMouseEventHandler = () => {
+  const [isDragging, setIsDragging] = useState(false);
+  const [startX, setStartX] = useState(0);
+  const [scrollLeft, setScrollLeft] = useState(0);
+
+  const startDragging = (e: React.PointerEvent<HTMLDivElement>) => {
+    setIsDragging(true);
+    setStartX(e.pageX - e.currentTarget.offsetLeft);
+    setScrollLeft(e.currentTarget.scrollLeft);
+  };
+
+  const stopDragging = () => {
+    setIsDragging(false);
+  };
+
+  const whileDragging = (e: React.PointerEvent<HTMLDivElement>) => {
+    if (!isDragging) return;
+    e.preventDefault();
+    const x = e.pageX - e.currentTarget.offsetLeft;
+    const walk = (x - startX) * 1;
+    e.currentTarget.scrollLeft = scrollLeft - walk;
+  };
+
+  return { startDragging, stopDragging, whileDragging };
+};
